Rename connection to connection-name in add conn
