+++ src/snowcli/cli/snowpark/procedure_coverage/report.py
@@ -11,7 +11,7 @@ from snowcli import config, utils
 from snowcli.config import connect_to_snowflake
 from snowcli.utils import generate_deploy_stage_name
 
-from snowcli.cli.snowpark import app
+from snowcli.cli.snowpark.procedure_coverage import app
 from snowcli.cli.common.flags import ConnectionOption
 
 log = logging.getLogger(__name__)
